fix: Remove iam group, add tailored iam user policy
Although IAM group is a tidy way of assigning permission to users needing similar access, it can be a security flaw in our application. IAM policy needs to be specific to individual user to prevent them from connecting to the rds instance as another user; specifically, it stops users from calling `generate-db-auth-token` of another known user and use their token to connect to the bde analytics instance.