feat: Create schema for each user
The purpose of this stack is to allow user to query bde data via fdw (remote to avoid tempering with production db). For this to be useful, user should be able to run some queries and save their changes locally. This commit creates a schema for each user upon user / role creation. It provides user with liberal privileges within their own schema.